+++ silence/utils/silence_json_encoder.py
@@ -1,7 +1,6 @@
 # Thanks to https://stackoverflow.com/a/3885198/5604339 :)
 import decimal
 from flask import json
-from datetime import datetime
 
 from silence.settings import settings
 
@@ -10,8 +9,15 @@ class SilenceJSONEncoder(json.JSONEncoder):
         if isinstance(o, decimal.Decimal):
             func = str if settings.DECIMALS_AS_STRINGS else float
             return func(o)
-        elif isinstance(o, datetime):
-            return o.isoformat()
+        elif hasattr(o, "isoformat"):
+            # Dates, times and datetimes should be formatted following ISO-8601
+            # with the format YYYY-MM-DD HH:MM:SS (or only the corresponding fragment
+            # for dates and times)
+            # ISO-8601 also defines "T" as a separator between date and time, which
+            # we replace by a space for compatibility with MySQL
+            # Instead of checking directly for their types, we test whether they
+            # have the isoformat() method, which all date/time classes share
+            return o.isoformat().replace("T", " ")
 
         try:
             return super().default(o)