+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 2.1.7-dev
-- ...
+- Fixes and improvements to the serialization of date/time objects to ensure compatibility with MySQL.
 
 # 2.1.7
 - Some improvements to logging: